\name{vardpoor-package}
\alias{vardpoor-package}
\alias{vardpoor}
\docType{package}
\title{
Estimation of indicators on social exclusion and poverty and its linearization, variance estimation
}
\description{Split data variables by domains, Weighted income percentile, linearization of the ratio estimator,
	     linearization of the at-risk-of-poverty rate, linearization of the at-risk-of-poverty threshold,
             linearization of the GINI coefficient, linearization of the gender pay (wage) gap, linearization
             of the median income below the at-risk-of-poverty gap, linearization of the income quintile share
             ratio, linearization of the relative median at-risk-of-poverty gap, residual estimation of
             calibration, the estimation of the simple random sampling, estimation of the variance and deff
             in domain, estimation of the variance using ultimate cluster method, estimation of the variance
             and deff for indicators on social exclusion and poverty.}
\details{
\tabular{ll}{
Package: \tab vardpoor\cr
Type: \tab Package\cr
Version: \tab 0.2.0.6\cr
Date: \tab 2014-03-12\cr
Depends: \tab R (>= 3.0.0), boot, MASS, foreach, laeken, lpSolve, data.table \cr
License: \tab GPL (>= 2)\cr
}

Index:
\preformatted{
domain                  Split data variable by domains
incPercentile           Weighted income percentile
lin.ratio               Linerazation of the ratio estimator
linarpr                 Linearization of at-risk-of-poverty rate
linarpt                 Linearization of at-risk-of-poverty threshold
lingini                 Linearization of the GINI coefficient I
lingini2                Linearization of the GINI coefficient II
lingpg                  Linearization of gender pay (wage) gap.
linpoormed              Linearization of the median income below
                        the at-risk-of-poverty gap
linqsr                  Linearization of quintile share ratio
linrmpg                 Linearization of relative median
                        at-risk-of-poverty gap
residual_est            Computes the estimation of residuals
var_srs                 Estimation of the simple random sampling
vardom                  Estimation of the variance in domain
vardomh                 Estimation of the variance in domain for household surveys
vardpoor-package        Estimation of indicators on social exclusion
                        and poverty and its linearization
variance_est            Estimation of the variance using ultimate cluster method
variance_othstr         Estimation of the variance using the new stratification
varpoord                Estimation of the variance for indicators on
                        social exclusion and poverty
}

}
\author{
  Juris Breidaks [aut, cre],
  Martins Liberts [aut]

Maintainer: Juris Breidaks <Juris.Breidaks@csb.gov.lv>
}

\keyword{ package }
